refactor(notifications): use Text component in all notifications
