+++ src/components/notifications/friendshipOffer.tsx
@@ -1,6 +1,7 @@
 import React from 'react'
 
-import { Container, Text } from './utils'
+import { Container } from './utils'
+import Text from '../text'
 
 import { useName } from '../../hooks/names'
 
@@ -30,7 +31,9 @@ export default function FriendshipOffer ({ data, onAccept, onDecline, onClose }:
   const name = useName(data.fromId)
 
   return <Container title={`${name} has offered you friendship.`}>
-    <Text text={data.text} />
+    <p>
+      <Text text={data.text} multiline />
+    </p>
 
     <div className={styles.ButtonsRow}>
       <button className={formStyles.OkButton} onClick={onAcceptFriendship}>
